requirements.txt
flask
requests
pandas
gunicorn
python-dotenv